refactor(native): group runtime interaction modules and tests
