+++ src/native/engine/RESOURCE_COLLECTION_INTEGRATION_GUIDE.md
@@ -205,11 +205,11 @@ When connecting to TypeScript API layer:
 ## Related Files
 
 - **Implementation**: `src/native/engine/runtime/player/player_resources.{h,c}`
-- **Implementation**: `src/native/engine/runtime/interactive_collection.{h,c}`
-- **Implementation**: `src/native/engine/runtime/interactive_object_spawner.{h,c}`
+- **Implementation**: `src/native/engine/runtime/interaction/interactive_collection.{h,c}`
+- **Implementation**: `src/native/engine/runtime/interaction/interactive_object_spawner.{h,c}`
 - **Tests**: `tests/native/player_resources_test.c` (28 tests)
-- **Tests**: `tests/native/interactive_collection_test.c` (26 tests)
-- **Tests**: `tests/native/interactive_object_spawner_test.c` (14 tests)
+- **Tests**: `tests/native/runtime/interaction/interactive_collection_test.c` (26 tests)
+- **Tests**: `tests/native/runtime/interaction/interactive_object_spawner_test.c` (14 tests)
 - **Existing Systems**: `src/native/engine/runtime/input/input_binding.{h,c}`
 - **Existing Systems**: `src/native/engine/runtime/terrain/terrain_chunks.{h,c}`
