+++ src/native/engine/runtime/interaction/interactive_objects.c
